+++ CHANGELOG.md
@@ -6,9 +6,12 @@ All notable changes to `Kronika` will be documented in this file.
 
 ### Changed
 
+### Fixed
+- Microsecond in `LocalDateTime::instant()`.
+
 ## [0.2.7](https://github.com/frostealth/kronika/releases/tag/0.2.7) – 2026-01-02
 ### Fixed
-- Formatter tests
+- Formatter tests.
 
 ## [0.2.6](https://github.com/frostealth/kronika/releases/tag/0.2.6) – 2025-11-19
 ### Added